+++ lib/src/main/cpp/CMakeLists.txt
@@ -1,10 +1,10 @@
 cmake_minimum_required(VERSION 3.16)
 
-project("sgbm_android_core")
+project("sgbm_android")
 
 option(BUILD_EXAMPLES "Build examples" OFF)
 
-add_subdirectory(core)
+add_subdirectory(src)
 if (BUILD_EXAMPLES)
     add_subdirectory(examples)
 endif ()
