+++ lib/src/main/cpp/src/conversions.cpp
