+++ lib/src/main/cpp/src/conversions.h
