+++ lib/src/main/cpp/src/logging.h
